Move tests in dashboard into appropriate folders
