--- src/legacy/core_plugins/kibana/public/dashboard/__tests__/get_app_state_mock.ts
@@ -1,48 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { AppStateClass } from '../legacy_imports';
-
-/**
- * A poor excuse for a mock just to get some basic tests to run in jest without requiring the injector.
- * This could be improved if we extract the appState and state classes externally of their angular providers.
- * @return {AppStateMock}
- */
-export function getAppStateMock(): AppStateClass {
-  class AppStateMock {
-    constructor(defaults: any) {
-      Object.assign(this, defaults);
-    }
-
-    on() {}
-    off() {}
-    toJSON() {
-      return '';
-    }
-    save() {}
-    translateHashToRison(stateHashOrRison: string | string[]) {
-      return stateHashOrRison;
-    }
-    getQueryParamName() {
-      return '';
-    }
-  }
-
-  return AppStateMock;
-}
