+++ src/legacy/core_plugins/kibana/public/dashboard/np_ready/dashboard_empty_screen.test.tsx
@@ -18,10 +18,7 @@
  */
 import React from 'react';
 import { mountWithIntl } from 'test_utils/enzyme_helpers';
-import {
-  DashboardEmptyScreen,
-  DashboardEmptyScreenProps,
-} from '../np_ready/dashboard_empty_screen';
+import { DashboardEmptyScreen, DashboardEmptyScreenProps } from './dashboard_empty_screen';
 // @ts-ignore
 import { findTestSubject } from '@elastic/eui/lib/test';
 import { coreMock } from '../../../../../../core/public/mocks';
