+++ src/legacy/core_plugins/kibana/public/dashboard/np_ready/dashboard_state.test.ts
@@ -19,7 +19,7 @@
 
 import './np_core.test.mocks';
 import { DashboardStateManager } from './dashboard_state_manager';
-import { getAppStateMock, getSavedDashboardMock } from '../__tests__';
+import { getAppStateMock, getSavedDashboardMock } from './test_utils';
 import { AppStateClass } from '../legacy_imports';
 import { DashboardAppState } from './types';
 import { TimeRange, TimefilterContract, InputTimeRange } from 'src/plugins/data/public';
